Organizing
Moving functions definitions into classes and cleaning up the main file. Started work on ability class.
