+++ Ability.py
@@ -1,0 +1,6 @@
+import main
+class Ability:
+    def __init__(self, name, target):
+        self.name = name
+        self.effect = []
+        self.magnitude = []
