+++ UI.py
@@ -1,0 +1,12 @@
+import os
+def wait():
+    input("\n\nPress enter to continue")
+
+def clear():
+        os.system('cls')
+
+class Input:
+    def __init__(self):
+        pass
+
+
